requirements.txt
# Function dependencies, for example:
# package>=version
tweepy==3.10.0
python-twitch-client
pandas
pathlib
pandas-gbq
pytrends
google-auth
gspread
oauth2client